+++ spinner/spinner_test.go
@@ -45,4 +45,7 @@ func (s *SpinnerSuite) TestSpinner(c *C) {
 	DisableAnimation = true
 	Show("ABCD")
 	Done(true)
+
+	Skip()
+	stopSpinner(_ACTION_DONE)
 }