+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `[spinner]` Added symbols customization
 * `[spinner]` Change default skip symbol to check mark
 * `[spinner]` Change default skip symbol color to dark grey
+* `[spinner]` Code refactoring
 
 ### 12.68.0